backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-text-splitters>=1.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]
